fix ghc-prim handling in booter
ghc-prim now needs to be installed before stage 1a, since an extra
step is required to expose the GHC.Prim module. the earlier fix
did this by hardcoding the package as an extra step. however, the
package was not added to the packages list, so ghc-prim was not
patched in the preparePackages step. this fix adds ghc-prim to the
boot.yaml config file again and adds it to the list of packages to
be prepared.